requirements.txt
langchain
llama-index
faiss-cpu
chromadb
openai
streamlit
PyYAML
pytest
click 